requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
pydantic[email]
python-dotenv
python-telegram-bot
httpx
